requirements.txt
# Google Services
google-auth>=2.36.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.160.0
gspread>=6.1.0

# Telegram + GitHub API calls
requests>=2.31.0

# Data Processing (update_sheet.py)
pandas>=2.2.0

# Scheduler
apscheduler>=3.10.0

# Utils
python-dotenv>=1.0.0
